feat: run local tests against k8s

+++ yarn-project/end-to-end/scripts/network_test.sh
@@ -32,6 +32,10 @@ FRESH_INSTALL="${FRESH_INSTALL:-false}"
 AZTEC_DOCKER_TAG=${AZTEC_DOCKER_TAG:-$(git rev-parse HEAD)}
 INSTALL_TIMEOUT=${INSTALL_TIMEOUT:-30m}
 CLEANUP_CLUSTER=${CLEANUP_CLUSTER:-false}
+DOCKER_TEST=${DOCKER_TEST:-true}
+
+export K8S="local"
+export INSTANCE_NAME="spartan"
 
 # Check required environment variable
 if [ -z "${NAMESPACE:-}" ]; then
@@ -140,18 +144,24 @@ helm upgrade --install spartan "$REPO/spartan/aztec-network/" \
 
 kubectl wait pod -l app==pxe --for=condition=Ready -n "$NAMESPACE" --timeout=10m
 
+set +x
 # Find 3 free ports between 9000 and 10000
-FREE_PORTS=$(comm -23 <(seq 9000 10000 | sort) <(ss -Htan | awk '{print $4}' | cut -d':' -f2 | sort -u) | shuf | head -n 3)
+FREE_PORTS=$(comm -23 <(seq 9000 10000 | sort) <(ss -Htan | awk '{print $4}' | cut -d':' -f2 | sort -u) | shuf | head -n 4)
+HELM_VALUES=$(helm get values -n "$NAMESPACE" spartan --all -o json)
 
-# Extract the free ports from the list
-PXE_PORT=$(echo $FREE_PORTS | awk '{print $1}')
-ANVIL_PORT=$(echo $FREE_PORTS | awk '{print $2}')
-METRICS_PORT=$(echo $FREE_PORTS | awk '{print $3}')
+export HOST_PXE_PORT=$(echo $FREE_PORTS | awk '{print $1}')
+export CONTAINER_PXE_PORT=$(jq -r '.pxe.service.nodePort' <<< $HELM_VALUES)
 
-GRAFANA_PASSWORD=$(kubectl get secrets -n metrics metrics-grafana -o jsonpath='{.data.admin-password}' | base64 --decode)
+export HOST_ETHEREUM_PORT=$(echo $FREE_PORTS | awk '{print $2}')
+export CONTAINER_ETHEREUM_PORT=$(jq -r '.ethereum.service.port' <<< $HELM_VALUES)
 
-# Namespace variable (assuming it's set)
-NAMESPACE=${NAMESPACE:-default}
+export HOST_METRICS_PORT=$(echo $FREE_PORTS | awk '{print $3}')
+export CONTAINER_METRICS_PORT=80
+
+export GRAFANA_PASSWORD=$(kubectl get secrets -n metrics metrics-grafana -o jsonpath='{.data.admin-password}' | base64 --decode)
+
+export LOG_LEVEL=${LOG_LEVEL:-"debug"}
+set -x
 
 # If we are unable to apply network shaping, as we cannot change existing chaos configurations, then delete existing configurations and try again
 if ! handle_network_shaping; then
@@ -164,20 +174,23 @@ if ! handle_network_shaping; then
   fi
 fi
 
-docker run --rm --network=host \
-  -v ~/.kube:/root/.kube \
-  -e K8S=local \
-  -e INSTANCE_NAME="spartan" \
-  -e SPARTAN_DIR="/usr/src/spartan" \
-  -e NAMESPACE="$NAMESPACE" \
-  -e HOST_PXE_PORT=$PXE_PORT \
-  -e CONTAINER_PXE_PORT=8081 \
-  -e HOST_ETHEREUM_PORT=$ANVIL_PORT \
-  -e CONTAINER_ETHEREUM_PORT=8545 \
-  -e HOST_METRICS_PORT=$METRICS_PORT \
-  -e CONTAINER_METRICS_PORT=80 \
-  -e GRAFANA_PASSWORD=$GRAFANA_PASSWORD \
-  -e DEBUG=${DEBUG:-""} \
-  -e LOG_JSON=1 \
-  -e LOG_LEVEL=${LOG_LEVEL:-"verbose"} \
-  aztecprotocol/end-to-end:$AZTEC_DOCKER_TAG $TEST
+if [ "$DOCKER_TEST" = "true" ]; then
+  docker run --rm --network=host \
+    -v ~/.kube:/root/.kube \
+    -e K8S \
+    -e INSTANCE_NAME \
+    -e SPARTAN_DIR="/usr/src/spartan" \
+    -e NAMESPACE \
+    -e HOST_PXE_PORT \
+    -e CONTAINER_PXE_PORT \
+    -e HOST_ETHEREUM_PORT \
+    -e CONTAINER_ETHEREUM_PORT \
+    -e HOST_METRICS_PORT \
+    -e CONTAINER_METRICS_PORT \
+    -e GRAFANA_PASSWORD \
+    -e LOG_LEVEL \
+    aztecprotocol/end-to-end:$AZTEC_DOCKER_TAG $TEST
+else
+  export SPARTAN_DIR=$(realpath "$SCRIPT_DIR/../../../spartan")
+  NODE_OPTIONS="${NODE_OPTIONS:-""} --no-warnings --experimental-vm-modules" yarn jest $TEST
+fi
